fix wx_shape validator for complex nested cases and add single value validation
