+++ tests/asdf_tests/test_validators.py
@@ -2,11 +2,14 @@ from io import BytesIO
 
 import asdf
 import jsonschema
+import numpy as np
 import pandas as pd
 import pytest
+from asdf import ValidationError
 
-from weldx.asdf.extension import WeldxAsdfExtension, WeldxExtension
 from weldx.asdf.tags.weldx.debug.test_property_tag import PropertyTagTestClass
+from weldx.asdf.tags.weldx.debug.test_shape_validator import ShapeValidatorTestClass
+from weldx.asdf.validators import _custom_shape_validator
 
 from .utility import _write_read_buffer
 
@@ -24,5 +27,141 @@ def test_property_tag_validator():
 
     with pytest.raises(jsonschema.exceptions.ValidationError):
         test = PropertyTagTestClass(prop3="STRING")
+        tree = {"root_node": test}
+        _write_read_buffer(tree)
+
+
+def test_shape_validator_syntax():
+    """Test handling of custom shape validation syntax in Python."""
+
+    def val(list_test, list_expected):
+        """Add shape key to lists."""
+        try:
+            _custom_shape_validator({"shape": list_test}, list_expected)
+            return True
+        except ValidationError:
+            return False
+
+    # correct evaluation
+    assert val([3], [3])
+    assert val([2, 4, 5], [2, 4, 5])
+    assert val([1, 2, 3], ["..."])
+    assert val([1, 2], [1, 2, "..."])
+    assert val([1, 2], ["...", 1, 2])
+    assert val([1, 2, 3], [1, 2, None])
+    assert val([1, 2, 3], [None, 2, 3])
+    assert val([1], [1, "..."])
+    assert val([1, 2, 3, 4, 5], [1, "..."])
+    assert val([1, 2, 3, 4, 5], ["...", 4, 5])
+    assert val([1, 2], [1, 2, "(3)"])
+    assert val([2, 3], ["(1)", 2, 3])
+    assert val([1, 2, 3], ["(1)", 2, 3])
+    assert val([2, 3], ["(1~3)", 2, 3])
+    assert val([2, 2, 3], ["(1~3)", 2, 3])
+    assert val([1, 2, 3], [1, "1~3", 3])
+    assert val([1, 2, 3], [1, "1~", 3])
+    assert val([1, 2, 3], [1, "~3", 3])
+    assert val([1, 2, 3], [1, "~", 3])
+    assert val([1, 200, 3], [1, "~", 3])
+    assert val([1, 2, 3], [1, 2, "(~)"])
+    assert val([1, 2, 300], [1, 2, "(~)"])
+    # assert val([1, 2, 3], [1, "(n)", "..."])  # should this be allowed?
+    _custom_shape_validator(1.0, [1])
+
+    # shape mismatch
+    assert not val([2, 2, 3], [1, "..."])
+    assert not val([2, 2, 3], ["...", 1])
+    assert not val([1], [1, 2])
+    assert not val([1, 2], [1])
+    assert not val([1, 2], [3, 2])
+    assert not val([1], [1, "~"])
+    assert not val([1], ["~", 1])
+    assert not val([1, 2, 3], [1, 2, "(4)"])
+    assert not val([1, 2, 3], ["(2)", 2, 3])
+    assert not val([1, 2], [1, "4~8"])
+    assert not val([1, 9], [1, "4~8"])
+    assert not val([1, 2], [1, "(4~8)"])
+    assert not val([1, 9], [1, "(4~8)"])
+
+    # syntax errors, these should throw a ValueError
+    with pytest.raises(ValueError):
+        val([1, 2], [1, "~", "(...)"])  # value error?
+    with pytest.raises(ValueError):
+        val([1, 2], [1, "(2)", 3])
+    with pytest.raises(ValueError):
+        val([1, 2], [1, 2, "((3))"])
+    with pytest.raises(ValueError):
+        val([1, 2], [1, 2, "3)"])
+    with pytest.raises(ValueError):
+        val([1, 2], [1, 2, "*3"])
+    with pytest.raises(ValueError):
+        val([1, 2], [1, 2, "(3"])
+    with pytest.raises(ValueError):
+        val([1, 2], [1, 2, "(3)3"])
+    with pytest.raises(ValueError):
+        val([1, 2], [1, 2, "2(3)"])
+    with pytest.raises(ValueError):
+        val([1, 2], [1, "...", 2])  # should this be allowed? syntax/value error?
+    with pytest.raises(ValueError):
+        val([1, 2], ["(1)", "..."])
+    with pytest.raises(ValueError):
+        val([1, 2], [1, "4~1"])
+    # no negative shape numbers allowed in syntax
+    with pytest.raises(ValueError):
+        val([-1, -2], [-1, -2])
+    with pytest.raises(ValueError):
+        val([1, 2], [1, 2, "(-3)"])
+    with pytest.raises(ValueError):
+        val([1, 2], [1, 2, "(-3~-1)"])
+    with pytest.raises(ValueError):
+        val([1, 2], [1, 2, "(-3~1)"])
+    with pytest.raises(ValueError):
+        val([1, 2, 1], ["(-3~1)", 2, 1])
+    with pytest.raises(ValidationError):  # test single value
+        _custom_shape_validator(1.0, [2])
+
+
+def test_shape_validator():
+    """Test custom ASDF shape validators."""
+    test = ShapeValidatorTestClass(
+        prop1=np.ones((1, 2, 3)),
+        prop2=np.ones((3, 2, 1)),
+        prop3=np.ones((2, 4, 6, 8, 10)),
+        prop4=np.ones((1, 3, 5, 7, 9)),
+        prop5=3.141,
+        nested_prop={"p1": np.ones((10, 8, 6, 4, 2)), "p2": np.ones((9, 7, 5, 3, 1))},
+    )
+    tree = {"root_node": test}
+    _write_read_buffer(tree)
+    # test_read = data["root_node"]
+    # TODO: add value assertion
+
+    with pytest.raises(ValidationError):
+        test = ShapeValidatorTestClass(
+            prop1=np.ones((1, 2, 3)),
+            prop2=np.ones((3, 2, 1)),
+            prop3=np.ones((2, 4, 6, 8, 10)),
+            prop4=np.ones((2, 3, 5, 7, 9)),  # mismatch a with prop5
+            prop5=3.141,
+            nested_prop={
+                "p1": np.ones((10, 8, 6, 4, 2)),
+                "p2": np.ones((9, 7, 5, 3, 1)),
+            },
+        )
+        tree = {"root_node": test}
+        _write_read_buffer(tree)
+
+    with pytest.raises(ValidationError):
+        test = ShapeValidatorTestClass(
+            prop1=np.ones((1, 2, 3)),
+            prop2=np.ones((5, 2, 1)),  # mismatch n with prop1
+            prop3=np.ones((2, 4, 6, 8, 10)),
+            prop4=np.ones((1, 3, 5, 7, 9)),
+            prop5=3.141,
+            nested_prop={
+                "p1": np.ones((10, 8, 6, 4, 2)),
+                "p2": np.ones((9, 7, 5, 3, 1)),
+            },
+        )
         tree = {"root_node": test}
         _write_read_buffer(tree)
