+++ tests/test_asdf.py
@@ -10,10 +10,7 @@ import pandas as pd
 import pytest
 from asdf import ValidationError
 
-from .asdf_tests.utility import _write_read_buffer
-
 from weldx.asdf.extension import WeldxAsdfExtension, WeldxExtension
-
 # weld design -----------------------------------------------------------------
 from weldx.asdf.tags.weldx.aws.design.base_metal import BaseMetal
 from weldx.asdf.tags.weldx.aws.design.connection import Connection
@@ -22,7 +19,6 @@ from weldx.asdf.tags.weldx.aws.design.sub_assembly import SubAssembly
 from weldx.asdf.tags.weldx.aws.design.weld_details import WeldDetails
 from weldx.asdf.tags.weldx.aws.design.weldment import Weldment
 from weldx.asdf.tags.weldx.aws.design.workpiece import Workpiece
-
 # welding process -----------------------------------------------------------------
 from weldx.asdf.tags.weldx.aws.process.arc_welding_process import ArcWeldingProcess
 from weldx.asdf.tags.weldx.aws.process.gas_component import GasComponent
@@ -30,15 +26,13 @@ from weldx.asdf.tags.weldx.aws.process.shielding_gas_for_procedure import (
     ShieldingGasForProcedure,
 )
 from weldx.asdf.tags.weldx.aws.process.shielding_gas_type import ShieldingGasType
-
 # iso groove -----------------------------------------------------------------
 from weldx.asdf.tags.weldx.core.iso_groove import get_groove
-
 # validators -----------------------------------------------------------------
-from weldx.asdf.tags.weldx.debug.test_shape_validator import ShapeValidatorTestClass
 from weldx.asdf.tags.weldx.debug.validator_testclass import ValidatorTestClass
-from weldx.asdf.validators import _custom_shape_validator
 from weldx.constants import WELDX_QUANTITY as Q_
+
+from .asdf_tests.utility import _write_read_buffer
 
 
 def test_aws_example():
@@ -352,107 +346,3 @@ def test_validators():
         )
         tree = {"root_node": test}
         data = _write_read_buffer(tree)
-
-
-def test_shape_validators():
-    """Test custom ASDF shape validators."""
-    test = ShapeValidatorTestClass(
-        prop1=np.ones((1, 2, 3)),
-        prop2=np.ones((3, 2, 1)),
-        prop3=np.ones((2, 4, 6, 8, 10)),
-        prop4=np.ones((1, 3, 5, 7, 9)),
-        nested_prop={"p1": np.ones((10, 8, 6, 4, 2)), "p2": np.ones((9, 7, 5, 3, 1))},
-    )
-
-    tree = {"root_node": test}
-
-    _write_read_buffer(tree)
-    # test_read = data["root_node"]
-    # TODO: add value assertion
-
-
-def test_shape_validator_syntax():
-    """Test handling of custom shape validation syntax in Python."""
-
-    def val(list_test, list_expected):
-        """Add shape key to lists."""
-        try:
-            _custom_shape_validator({"shape": list_test}, list_expected)
-            return True
-        except ValidationError:
-            return False
-
-    # correct evaluation
-    assert val([3], [3])
-    assert val([2, 4, 5], [2, 4, 5])
-    assert val([1, 2, 3], ["..."])
-    assert val([1, 2], [1, 2, "..."])
-    assert val([1, 2], ["...", 1, 2])
-    assert val([1, 2, 3], [1, 2, None])
-    assert val([1, 2, 3], [None, 2, 3])
-    assert val([1], [1, "..."])
-    assert val([1, 2, 3, 4, 5], [1, "..."])
-    assert val([1, 2, 3, 4, 5], ["...", 4, 5])
-    assert val([1, 2], [1, 2, "(3)"])
-    assert val([2, 3], ["(1)", 2, 3])
-    assert val([1, 2, 3], ["(1)", 2, 3])
-    assert val([2, 3], ["(1~3)", 2, 3])
-    assert val([2, 2, 3], ["(1~3)", 2, 3])
-    assert val([1, 2, 3], [1, "1~3", 3])
-    assert val([1, 2, 3], [1, "1~", 3])
-    assert val([1, 2, 3], [1, "~3", 3])
-    assert val([1, 2, 3], [1, "~", 3])
-    assert val([1, 200, 3], [1, "~", 3])
-    assert val([1, 2, 3], [1, 2, "(~)"])
-    assert val([1, 2, 300], [1, 2, "(~)"])
-    # assert val([1, 2, 3], [1, "(n)", "..."])  # should this be allowed?
-
-    # shape mismatch
-    assert not val([2, 2, 3], [1, "..."])
-    assert not val([2, 2, 3], ["...", 1])
-    assert not val([1], [1, 2])
-    assert not val([1, 2], [1])
-    assert not val([1, 2], [3, 2])
-    assert not val([1], [1, "~"])
-    assert not val([1], ["~", 1])
-    assert not val([1, 2, 3], [1, 2, "(4)"])
-    assert not val([1, 2, 3], ["(2)", 2, 3])
-    assert not val([1, 2], [1, "4~8"])
-    assert not val([1, 9], [1, "4~8"])
-    assert not val([1, 2], [1, "(4~8)"])
-    assert not val([1, 9], [1, "(4~8)"])
-
-    # syntax errors, these should throw a ValueError
-    with pytest.raises(ValueError):
-        val([1, 2], [1, "~", "(...)"])  # value error?
-    with pytest.raises(ValueError):
-        val([1, 2], [1, "(2)", 3])
-    with pytest.raises(ValueError):
-        val([1, 2], [1, 2, "((3))"])
-    with pytest.raises(ValueError):
-        val([1, 2], [1, 2, "3)"])
-    with pytest.raises(ValueError):
-        val([1, 2], [1, 2, "*3"])
-    with pytest.raises(ValueError):
-        val([1, 2], [1, 2, "(3"])
-    with pytest.raises(ValueError):
-        val([1, 2], [1, 2, "(3)3"])
-    with pytest.raises(ValueError):
-        val([1, 2], [1, 2, "2(3)"])
-    with pytest.raises(ValueError):
-        val([1, 2], [1, "...", 2])  # should this be allowed? syntax/value error?
-    with pytest.raises(ValueError):
-        val([1, 2], ["(1)", "..."])
-    with pytest.raises(ValueError):
-        val([1, 2], [1, "4~1"])
-    # no negative shape numbers allowed in syntax
-    with pytest.raises(ValueError):
-        val([-1, -2], [-1, -2])
-    with pytest.raises(ValueError):
-        val([1, 2], [1, 2, "(-3)"])
-    with pytest.raises(ValueError):
-        val([1, 2], [1, 2, "(-3~-1)"])
-    with pytest.raises(ValueError):
-        val([1, 2], [1, 2, "(-3~1)"])
-    with pytest.raises(ValueError):
-        val([1, 2, 1], ["(-3~1)", 2, 1])
