+++ weldx/asdf/schemas/weldx.bam.de/weldx/debug/validator_testclass-1.0.0.yaml
@@ -5,7 +5,7 @@ id: "http://weldx.bam.de/schemas/weldx/debug/validator_testclass-1.0.0"
 tag: "tag:weldx.bam.de:weldx/debug/validator_testclass-1.0.0"
 
 title: |
-  simple demonstration schema for wx_unit and wx_shape validator
+  simple demonstration schema for wx_unit validator
 type: object
 properties:
   length_prop:
@@ -40,9 +40,6 @@ properties:
         $ref: tag:stsci.edu:asdf/unit/quantity-1.1.0
     wx_unit:
       q1: "m"
-    wx_shape:
-      q1:
-        value: [3,3]
 
   simple_prop:
     description: simple property without any references ref
@@ -64,7 +61,4 @@ wx_unit:
   current_prop: A
   nested_prop:
     q2: "m*mm*cm"
-wx_shape:
-  current_prop:
-    value: [2,2]
 ...
