+++ weldx/asdf/validators.py
@@ -328,36 +328,43 @@ def _custom_shape_validator(dict_test, dict_expected):
     # if dict_test.keys() != dict_expected.keys():
     #     return False
 
+    dict_values = {}
+
     # catch single shape definitions
     if isinstance(dict_expected, list):
-        if "shape" not in dict_test:
+        if isinstance(dict_test, (float, int)):  # test against [1] for single values
+            list_test, list_expected = _prepare_list([1], dict_expected)
+        elif "shape" in dict_test:
+            list_test, list_expected = _prepare_list(dict_test["shape"], dict_expected)
+        else:
             return ValidationError(f"Could not find shape key in instance {dict_test}.")
-        list_test, list_expected = _prepare_list(dict_test["shape"], dict_expected)
 
         _validate_expected_list(list_expected)
         _dict_values = _compare_lists(list_test, list_expected)
+
+        if _dict_values is False:
+            raise ValidationError(
+                f"Shape {list_test[::-1]} does not match requirement "
+                f"{list_expected[::-1]}"
+            )
+
+        return _dict_values
 
     elif isinstance(dict_expected, dict):
         for item in dict_expected:
             # go one level deeper in the dictionary
             _dict_values = _custom_shape_validator(dict_test[item], dict_expected[item])
+
+            for key in _dict_values:
+                if key not in dict_values:
+                    dict_values[key] = _dict_values[key]
+                elif dict_values[key] != _dict_values[key]:
+                    return False
     else:
         raise ValueError(
             f"Found an incorrect object: {type(dict_expected)}. "
             "Should be a dict or list."
         )
-
-    if _dict_values is False:
-        raise ValidationError(
-            f"Shape {list_test[::-1]} does not match requirement {list_expected[::-1]}"
-        )
-
-    dict_values = {}
-    for key in _dict_values:
-        if key not in dict_values:
-            dict_values[key] = _dict_values[key]
-        elif dict_values[key] != _dict_values[key]:
-            return False
 
     return dict_values
